+++ app/src/main/java/com/moviesearcher/rated/movie/viewmodel/RatedMoviesViewModel.kt
@@ -8,7 +8,7 @@ import com.moviesearcher.rated.movie.model.RatedMoviesResponse
 class RatedMoviesViewModel : ViewModel() {
     private lateinit var ratedMovies: LiveData<RatedMoviesResponse>
 
-    fun getRatedMovies(accountId: Int, sessionId: String): LiveData<RatedMoviesResponse> {
+    fun getRatedMovies(accountId: Long, sessionId: String): LiveData<RatedMoviesResponse> {
         ratedMovies = Api.getRatedMovies(accountId, sessionId)
 
         return ratedMovies
