+++ app/src/main/java/com/moviesearcher/rated/tvepisode/viewmodel/RatedTvEpisodesViewModel.kt
@@ -8,7 +8,7 @@ import com.moviesearcher.rated.tvepisode.model.RatedTvEpisodesResponse
 class RatedTvEpisodesViewModel : ViewModel() {
     private lateinit var ratedTvEpisodes: LiveData<RatedTvEpisodesResponse>
 
-    fun getRatedTvEpisodes(accountId: Int, sessionId: String): LiveData<RatedTvEpisodesResponse> {
+    fun getRatedTvEpisodes(accountId: Long, sessionId: String): LiveData<RatedTvEpisodesResponse> {
         ratedTvEpisodes = Api.getRatedTvEpisodes(accountId, sessionId)
 
         return ratedTvEpisodes
